% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.table.R
\docType{data}
\name{pos}
\alias{pos}
\title{Parts of speech for English words from the Moby Project.}
\format{
Data.table with 246690 rows and 3 variables
#' \describe{
  \item{word}{Lowercase English word or phrase}
  \item{pos}{Lowercase English part of speech, one per row}
  \item{multi_word}{TRUE if the word record has a space (contains multiple words), else FALSE.}
}
}
\source{
\url{https://archive.org/details/mobypartofspeech03203gut}
}
\usage{
pos
}
\description{
Parts of speech for English words/phrases from the Moby Project by Grady Ward.
Words with non-ASCII characters have been removed. One row per word + part of speech
}
\keyword{datasets}
