% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.table.R
\docType{data}
\name{regex_paragraph}
\alias{regex_paragraph}
\title{Regular expression that might be used to split strings of text into component paragraphs.}
\format{
A string
}
\usage{
regex_paragraph
}
\description{
"\\n", A regular expression to split strings when encountering a new line.
}
\keyword{datasets}
