% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.table.R
\docType{data}
\name{regex_sentence}
\alias{regex_sentence}
\title{Regular expression that might be used to split strings of text into component sentences.}
\format{
A string
}
\usage{
regex_sentence
}
\description{
"[.?!]\\s", A regular expression to split strings when encountering a common end of sentence punctuation pattern.
}
\keyword{datasets}
