% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.table.R
\docType{data}
\name{stopwords}
\alias{stopwords}
\title{Vector of lowercase English stop words.}
\format{
A vector of 728 unique English stop words in lowercase
}
\source{
\url{http://snowball.tartarus.org/algorithms/english/stop.txt}

\url{http://www.lextek.com/manuals/onix/stopwords1.html}

\url{http://www.lextek.com/manuals/onix/stopwords2.html}
}
\usage{
stopwords
}
\description{
Unique lowercase English stop words from 3 lexicons combined into one vector.
Combines snowball, onix, and SMART lists of stopwords.
}
\keyword{datasets}
