% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.table.R
\name{str_any_match}
\alias{str_any_match}
\title{Detect if there are any words in a vector also found in another vector.}
\usage{
str_any_match(x, y)
}
\arguments{
\item{x}{A vector of words.}

\item{y}{A vector of words to test against.}
}
\value{
TRUE/FALSE, TRUE if any words in x are also in y
}
\description{
Detect if there are any words in a vector also found in another vector.
}
\examples{
str_any_match(
x = c("a", "dog", "went", "to", "the", "store"),
y = c("the")
)
str_any_match(
x = c("a", "dog", "went", "to", "the", "store"),
y = c("apple")
)
}
