% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.table.R
\name{str_count_intersect}
\alias{str_count_intersect}
\title{Count the intersecting words in a vector that are found in another vector (only counts unique words).}
\usage{
str_count_intersect(x, y)
}
\arguments{
\item{x}{A vector of words.}

\item{y}{A vector of words to test against.}
}
\value{
A number, the count of unique words in x also in y
}
\description{
Count the intersecting words in a vector that are found in another vector (only counts unique words).
}
\examples{
str_count_intersect(
x = c("a", "dog", "went", "to", "the", "store"),
y = c("dog", "to", "store")
)
}
