% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.table.R
\name{str_count_jaccard_similarity}
\alias{str_count_jaccard_similarity}
\title{Calculates the intersect divided by union of two vectors of words.}
\usage{
str_count_jaccard_similarity(x, y)
}
\arguments{
\item{x}{A vector of words.}

\item{y}{A vector of words to test against.}
}
\value{
A number, the intersect divided by union of two vectors of words.
}
\description{
Calculates the intersect divided by union of two vectors of words.
}
\examples{

str_count_jaccard_similarity(
x = c("a", "dog", "went", "to", "the", "store"),
y = c("this", "dog", "went", "to", "the", "store")
)
}
