% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.table.R
\name{str_count_positional_match}
\alias{str_count_positional_match}
\title{Count words from a vector that are found in the same position in another vector.}
\usage{
str_count_positional_match(x, y, ratio = FALSE)
}
\arguments{
\item{x}{A vector of words.}

\item{y}{A vector of words to test against.}

\item{ratio}{TRUE/FALSE, if TRUE, returns the number of words in x with a positional match in y divided by the number of words in x.}
}
\value{
A count of the words in x with matches in the same position in y.
}
\description{
Count words from a vector that are found in the same position in another vector.
}
\examples{
str_count_positional_match(
x = c("a", "dog", "went", "to", "the", "store"),
y = c("this", "dog", "ran", "from", "the", "store")
)
}
