% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.table.R
\name{str_count_setdiff}
\alias{str_count_setdiff}
\title{Count the words in a vector that don't intersect with another vector (only counts unique words).}
\usage{
str_count_setdiff(x, y)
}
\arguments{
\item{x}{A vector of words.}

\item{y}{A vector of words to test against.}
}
\value{
A number, the count of unique words in x not also in y
}
\description{
Count the words in a vector that don't intersect with another vector (only counts unique words).
}
\examples{
str_count_setdiff(
x = c("a", "dog", "dog", "went", "to", "the", "store"),
y = c("dog", "to", "store")
)
}
