% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.table.R
\name{str_counts}
\alias{str_counts}
\title{Create a list of a vector of unique words found in x and a vector of the counts of each word in x.}
\usage{
str_counts(x)
}
\arguments{
\item{x}{A vector of words.}
}
\value{
A list, position one is a vector of unique and sorted words from x, position two is a vector of the counts for each word.
}
\description{
Create a list of a vector of unique words found in x and a vector of the counts of each word in x.
}
\examples{
str_counts(
x = c("a", "dog", "went", "to", "the", "store", "and", "a", "dog", "went", "to", "another", "store")
)
}
