% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.table.R
\name{str_extract_positional_match}
\alias{str_extract_positional_match}
\title{Extract words from a vector that are found in the same position in another vector.}
\usage{
str_extract_positional_match(x, y)
}
\arguments{
\item{x}{A vector of words.}

\item{y}{A vector of words to test against.}
}
\value{
x, with the words without matches in the same position in y removed.
}
\description{
Extract words from a vector that are found in the same position in another vector.
}
\examples{
str_extract_positional_match(
x = c("a", "dog", "went", "to", "the", "store"),
y = c("this", "dog", "ran", "from", "the", "store")
)
}
