% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.table.R
\name{str_extract_positional_nomatch}
\alias{str_extract_positional_nomatch}
\title{Extract words from a vector that are not found in the same position in another vector.}
\usage{
str_extract_positional_nomatch(x, y)
}
\arguments{
\item{x}{A vector of words.}

\item{y}{A vector of words to test against.}
}
\value{
x, with the words with matches found in the same position in y removed.
}
\description{
Extract words from a vector that are not found in the same position in another vector.
}
\examples{
str_extract_positional_nomatch(
x = c("a", "dog", "went", "to", "the", "store"),
y = c("a", "crazy", "dog", "ran", "from", "the", "store")
)
}
