% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.table.R
\name{str_rm_blank_space}
\alias{str_rm_blank_space}
\title{Remove and replace excess white space from strings.}
\usage{
str_rm_blank_space(x, replacement = " ")
}
\arguments{
\item{x}{A vector or string.}

\item{replacement, }{A string to replace the blank space with, defaults to " ", which replaces excess space with a single space.}
}
\value{
x, with extra white space removed/replaced.
}
\description{
Remove and replace excess white space from strings.
}
\examples{
str_rm_blank_space(c("this     is   a test.  ", "    will it    work? "))
}
