% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.table.R
\name{str_rm_long_words}
\alias{str_rm_long_words}
\title{Remove words from a vector that have more than a maximum number of characters.}
\usage{
str_rm_long_words(x, max_char_length)
}
\arguments{
\item{x}{A vector of words.}

\item{max_char_length, }{An integer, the maximum number of characters a word can have to not be removed.}
}
\value{
x, with the words not having a character count less than or equal to the max_char_length removed.
}
\description{
Remove words from a vector that have more than a maximum number of characters.
}
\examples{
str_rm_long_words(
x = c("a", "dog", "went", "to", "the", "store"),
max_char_length = 2
)
}
