% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.table.R
\name{str_rm_non_alphanumeric}
\alias{str_rm_non_alphanumeric}
\title{Remove and replace non-alphanumeric characters from strings.}
\usage{
str_rm_non_alphanumeric(x, replacement = " ")
}
\arguments{
\item{x}{A vector or string.}

\item{replacement, }{A string to replace the numbers with, defaults to " ".}
}
\value{
x, with non-alphanumeric (A-z, 0-9) characters removed/replaced.
}
\description{
Remove and replace non-alphanumeric characters from strings.
}
\examples{
str_rm_non_alphanumeric(c("test 67890 * \% $ "))
}
