% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.table.R
\name{str_rm_non_printable}
\alias{str_rm_non_printable}
\title{Remove and replace non-printable characters from strings.}
\usage{
str_rm_non_printable(x, replacement = " ")
}
\arguments{
\item{x}{A vector or string.}

\item{replacement, }{A string to replace the numbers with, defaults to " ".}
}
\value{
x, with non-printable characters removed/replaced.
}
\description{
Remove and replace non-printable characters from strings.
}
\examples{
str_rm_non_printable(c("test \n\n\n67890 * \% $ "))
}
