% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.table.R
\name{str_rm_numbers}
\alias{str_rm_numbers}
\title{Remove and replace numbers from strings.}
\usage{
str_rm_numbers(x, replacement = "")
}
\arguments{
\item{x}{A vector or string.}

\item{replacement, }{A string to replace the numbers with, defaults to "".}
}
\value{
x, with numbers 0-9 removed/replaced.
}
\description{
Remove and replace numbers from strings.
}
\examples{
str_rm_numbers(c("test 1a234b5", "test 67890"))
}
