% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.table.R
\name{str_rm_regexp_match}
\alias{str_rm_regexp_match}
\title{Remove words from a vector that match a regular expression.}
\usage{
str_rm_regexp_match(x, pattern)
}
\arguments{
\item{x}{A vector of words.}

\item{pattern, }{A regular expression.}
}
\value{
x, with the words matching the regular expression removed.
}
\description{
Remove words from a vector that match a regular expression.
}
\examples{
str_rm_regexp_match(
x = c("a", "dog", "went", "to", "the", "store"),
pattern = "to"
)
}
