% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.table.R
\name{str_rm_short_words}
\alias{str_rm_short_words}
\title{Remove words from a vector that don't have a minimum number of characters.}
\usage{
str_rm_short_words(x, min_char_length)
}
\arguments{
\item{x}{A vector of words.}

\item{min_char_length, }{An integer, the minimum number of characters a word can have to not be removed.}
}
\value{
x, with the words not having a character count greater than or equal to the min_char_length removed.
}
\description{
Remove words from a vector that don't have a minimum number of characters.
}
\examples{
str_rm_short_words(
x = c("a", "dog", "went", "to", "the", "store"),
min_char_length = 3
)
}
