% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.table.R
\name{str_rm_words}
\alias{str_rm_words}
\title{Remove words from a vector of words found in another vector of words.}
\usage{
str_rm_words(x, y = stopwords)
}
\arguments{
\item{x}{A vector of words.}

\item{y, }{A vector of words to delete from x, defaults to English stop words.}
}
\value{
x, with the words found in y removed.
}
\description{
Remove words from a vector of words found in another vector of words.
}
\examples{
str_rm_words(
x = c("a", "dog", "went", "to", "the", "store"),
y = stopwords
)

str_rm_words(
x = c("a", "dog", "went", "to", "the", "store"),
y = c("dog", "store")
)
}
