% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.table.R
\name{str_tolower}
\alias{str_tolower}
\title{Calls base::tolower(), which converts letters to lowercase. Only included to point
out that base::tolower exists and should be used directly.}
\usage{
str_tolower(x)
}
\arguments{
\item{x}{A vector or string.}
}
\value{
x, converted to lowercase.
}
\description{
Calls base::tolower(), which converts letters to lowercase. Only included to point
out that base::tolower exists and should be used directly.
}
\examples{
str_tolower(c("ALLCAPS", "Some capS"))
}
