% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.table.R
\name{str_weighted_count_match}
\alias{str_weighted_count_match}
\title{Weighted count of the words in a vector that are found in another vector.}
\usage{
str_weighted_count_match(x, y)
}
\arguments{
\item{x}{A list of words and counts created by str_counts(x).}

\item{y}{A list of words and counts created by str_counts(y).}
}
\value{
A number, the count of words in x also in y scaled by the number of times each word appears in x and y.
If a word appears 3 times in x and 2 times in y, the result is 6, assuming no other words match.
}
\description{
Weighted count of the words in a vector that are found in another vector.
}
\examples{
str_weighted_count_match(
x = str_counts(c("a", "dog", "dog", "went", "to", "the", "store")),
y = str_counts(c("dog", "dog", "dog"))
)
}
