% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_address.R
\name{text.address.parser}
\alias{text.address.parser}
\title{text.address.parser}
\usage{
\method{text}{address.parser}(address, dict_fixed = NULL,
  dict_regexp = NULL, dict_regexp_number = NULL, verbose = FALSE)
}

