% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_clean.R
\name{text.clean}
\alias{text.clean}
\title{text.clean}
\usage{
text.clean(
  x,
  trim = TRUE,
  remove_links = TRUE,
  remove_punctuation = TRUE,
  remove_numbers = TRUE,
  remove_tabs = TRUE,
  remove_newlines = TRUE,
  remove_extra_spaces = TRUE)
}

