% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_do.R
\name{text.DigOrig.AddressNormalizer}
\alias{text.DigOrig.AddressNormalizer}
\title{text.DigOrig.AddressNormalizer}
\usage{
text.DigOrig.AddressNormalizer(
 input_address,
 input_city,
 input_zipcode,
 ip_PuntuaMiZona = connData$puntuamizona_ip,
 user_PuntuaMiZona = connData$puntuamizona_user,
 pass_PuntuaMiZona = connData$puntuamizona_pass)
}

