% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_do.R
\name{text.DigOrig.Normalizer}
\alias{text.DigOrig.Normalizer}
\title{text.DigOrig.Normalizer}
\usage{
\method{text}{DigOrig.Normalizer}(input_address, input_city, input_zipcode,
  ip_PuntuaMiZona, user_PuntuaMiZona, pass_PuntuaMiZona)
}

