% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_do.R
\name{text.DigOrig.PuntuaMiZona}
\alias{text.DigOrig.PuntuaMiZona}
\title{text.DigOrig.PuntuaMiZona}
\usage{
text.DigOrig.PuntuaMiZona(
 street,
 zipcode,
 ip_PuntuaMiZona = connData$puntuamizona_ip,
 user_PuntuaMiZona = connData$puntuamizona_user,
 pass_PuntuaMiZona = connData$puntuamizona_pass,
 throw_error = texttools:::param.default.text.PuntuaMiZona.ThrowError)
}

