% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{text.filter}
\alias{text.filter}
\title{text.filter}
\usage{
text.filter(
  x,
  pattern,
  ignore.case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  message = texttools:::param.default.text.filter.message,
  batch = texttools:::param.default.text.filter.batch)
}

