% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{text.set.default.text.PuntuaMiZona.ThrowError}
\alias{text.set.default.text.PuntuaMiZona.ThrowError}
\title{text.set.default.text.PuntuaMiZona.ThrowError}
\usage{
text.set.default.text.PuntuaMiZona.ThrowError(value)
}

